requirements.txt
beautifulsoup4==4.12.3
requests==2.32.2
nltk==3.8.1
scikit-learn==1.5.0
google-cloud-storage==2.16.0
fastapi==0.111.0
uvicorn==0.30.1
